配置db，mapper

+++ src/be/src/main/java/hello/service/User.java
@@ -1,19 +1,19 @@
 package hello.service;
 
 public class User {
-    Integer id;
+    Long id;
     String name;
 
-    public User(Integer id, String name) {
+    public User(Long id, String name) {
         this.id = id;
         this.name = name;
     }
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
